requirements.txt
pyyaml
pyserial

